Added confirm dialogue when deleting entries Removed duplicate validation checks from form classes and adjusted unique constraints to let the DB handle it Added error handling for resource maniputation on duplicate entries and other errors Adjusted flask flash messages to include an error category where appropriate + styling Changed book.series_index to book.series_number to better reflect its intended use. Array-style numbering doesn't make much sense here Added alembic migrations that correspond to unique constraint changes

+++ app/static/script.js
@@ -9,11 +9,15 @@ $(document).ready(function(){
     $('#delete').click(function(event){
         event.preventDefault();
 
-        url = window.location.href;
-        delete_url = url.substring(0, url.lastIndexOf('/'));
-
-        console.log("Deleting: " + delete_url);
-        deleteResource(delete_url);
+        if (confirm('Are you sure you want to delete this entry?')) {
+            url = window.location.href;
+            delete_url = url.substring(0, url.lastIndexOf('/'));
+    
+            console.log("Deleting: " + delete_url);
+            deleteResource(delete_url);
+        } else {
+            console.log('Deletion cancelled.');
+        }
     })
 
     function deleteResource(url) {
